requirements.txt
fastapi
python-dotenv
python-telegram-bot==20.7
python-dateutil
pytz
schedule
requests
uvicorn
